# swagger.yml
openapi: 3.0.0

info:
  title: Champions API
  version: 1.0.0
  description: API for user registration, login and protected routes with JWT authentication.

servers:
  - url: http://localhost:4000
    description: Dev server

paths:
  /register:
    post:
      tags:
        - "Auth"
      summary: Register a new user and obtain a JWT token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: The username of the new user
                password:
                  type: string
                  description: The password of the new user
              required:
                - username
                - password
      responses:
        '201':
          description: JWT token generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: The JWT token
        '400':
          description: Missing username or password
  /login:
    post:
      tags:
        - "Auth"
      summary: Login a user and obtain a JWT token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: The username of the user
                password:
                  type: string
                  description: The password of the user
              required:
                - username
                - password
      responses:
        '200':
          description: JWT token generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: The JWT token
        '401':
          description: Invalid credentials
  /video:
    post:
      tags:
        - "Video"
      summary: Send video
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                video:
                  type: string
                  format: binary
                  description: The uploaded file data
      responses:
        '200':
          description: Start generate
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
    get:
      tags:
        - "Video"
      summary: Get history of videos
      parameters:
      - name: id
        in: query
        schema:
          type: array
          items:
            type: integer
            format: int64
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Video list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      description: id of video
                    status:
                      type: string
                      description: Status of video
                    object_name:
                      type: string
                      description: Video
    delete:
      tags:
        - "Video"
      summary: Remove all images from history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Video Removed
  /files/{object_name}:
    get:
      tags:
        - "Files"
      summary: Get file from history
      parameters:
        - in: path
          name: object_name
          schema:
            type: string
          required: true
          description: object name
      security:
        - bearerAuth: []
      responses:
        '200':
          description: File
  /video/{id}:
    delete:
      tags:
        - "Video"
      summary: Remove video from history
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: id
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Video Removed
  /clip:
    get:
      tags:
        - "Clip"
      summary: Get history of Clip
      parameters:
      - name: video_id
        in: query
        schema:
          type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Clip list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      description: id of Clip
                    object_name:
                      type: string
                      description: Clip
  /clip/{id}:
    delete:
      tags:
        - "Clip"
      summary: Remove Clip from history
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: id
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Clip Removed
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Image:
      type: object
      properties:
        width:
          type: string
          description: Width of image
          example: 256
        height:
          type: string
          description: Height of image
          example: 256
        count:
          type: string
          description: Count of images to generate
          example: 1
        product_type:
          type: string
          description: Type of product to generate
          example: credit_card
        positive_prompt:
          type: string
          description: Positive promt
        negative_prompt:
          type: string
          description: Negative promt
        offer:
          type: string
          description: Offer
      required:
        - width
        - height
        - count
        - product_type